requirements.txt
Flask==3.0.0 #2.1.3
Flask-Cors==3.0.10
flask-openapi3==4.3.0
Flask-SQLAlchemy==2.5.1
nose2==0.12.0
pydantic==2.12.5
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
typing_extensions==4.15.0
werkzeug==3.1.4
flask-openapi3[swagger]
python-dotenv==1.2.1
google-genai==1.55.0